* remove group * move is_removable_check to util functions

+++ app/structure/views_delete.py
@@ -3,6 +3,8 @@ from flask_login import login_required
 
 from app import db
 from app.decorators import check_master
+from app.is_removable_check import is_citizenship_removable, is_parent_removable, is_school_removable, \
+    is_section_removable, is_group_removable
 from app.models import Citizenship, Section, Parent, School, Group
 from app.structure import structure
 
@@ -12,7 +14,7 @@ from app.structure import structure
 @check_master
 def delete_citizenship(id):
     citizenship = Citizenship.query.get_or_404(id)
-    if citizenship.students.count() > 0: abort(409)
+    if not is_citizenship_removable(citizenship): abort(409)
     db.session.delete(citizenship)
     flash('гражданство удалено')
     return redirect(url_for('.citizenships_list'))
@@ -23,7 +25,7 @@ def delete_citizenship(id):
 @check_master
 def delete_section(id):
     section = Section.query.get_or_404(id)
-    if section.groups.count() > 0: abort(409)
+    if not is_section_removable(section): abort(409)
     db.session.delete(section)
     flash('секция удалена')
     return redirect(url_for('.sections_list'))
@@ -34,7 +36,7 @@ def delete_section(id):
 @check_master
 def delete_school(id):
     school = School.query.get_or_404(id)
-    if school.students.count() > 0: abort(409)
+    if not is_school_removable(school): abort(409)
     db.session.delete(school)
     flash('школа удалена')
     return redirect(url_for('.schools_list'))
@@ -45,7 +47,7 @@ def delete_school(id):
 @check_master
 def delete_parent(id):
     parent = Parent.query.get_or_404(id)
-    if parent.parent_of_students.count() > 0: abort(409)
+    if not is_parent_removable(parent): abort(409)
     db.session.delete(parent)
     flash('родитель удален')
     return redirect(url_for('.parents_list'))
@@ -56,6 +58,7 @@ def delete_parent(id):
 @check_master
 def delete_group(id):
     group = Group.query.get_or_404(id)
+    if not is_group_removable(group): abort(409)
     db.session.delete(group)
-    flash('группа удалено')
+    flash('группа удалена')
     return redirect(url_for('.groups_list'))
